+++ packages/core/src/lib/field/field.ts
@@ -17,6 +17,8 @@ export class Field implements FieldConfigProperty {
   view?: string;
   /** Custom Validation function */
   validate?: (value, field) => any;
+  /** Custom clean function to prepare for save */
+  beforeSave?: (value?, field?, body?) => any;
   /** Custom Component to display form input **/
   input?: any;
   /** Custom Component to display value **/
@@ -79,5 +81,9 @@ export class Field implements FieldConfigProperty {
   /** Returns the view for the given occasion */
   getView(occasion?) {
     return this[occasion + 'View'] || this.view;
+  }
+  /** Returns the component for the given occasion */
+  getComponent(occasion?) {
+    return this[occasion + 'Component'];
   }
 }