+++ packages/ui/src/lib/actionbar/actionbar.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input, ElementRef, ChangeDetectorRef } from '@angular/core';
 import { SelectComponent } from '../select/select.component';
+import { selectTemplate } from '../select/select.component.html';
 import { ListConfig, List, Item } from '@ec.components/core';
 
 export type ActionFunction = (item?: Item<Action> | any, actionbar?: ActionbarComponent) => any;
@@ -17,9 +18,10 @@ export interface ActionbarConfig extends ListConfig<Action> {}
 
 @Component({
   selector: 'ec-actionbar',
-  templateUrl: '../select/select.component.html',
+  template: selectTemplate,
 })
 export class ActionbarComponent extends SelectComponent<Action> implements OnInit {
+  @Input() root = 'ROOT'; // id of root stack item
   @Input() config: ActionbarConfig = {
     label: 'title',
     identifier: 'id',
@@ -63,7 +65,7 @@ export class ActionbarComponent extends SelectComponent<Action> implements OnIni
 
   currentID() {
     if (!this.selection || this.selection.isEmpty()) {
-      return 'ROOT';
+      return this.root;
     }
     return this.selection.items[this.selection.items.length - 1].id();
   }
@@ -77,11 +79,17 @@ export class ActionbarComponent extends SelectComponent<Action> implements OnIni
     return this.actionStack[this.currentID()];
   }
 
-  loadActions(actions, addToStack = true) {
+  async loadActions(actions, addToStack = true) {
+    let resolved;
+    if (typeof actions === 'function') {
+      resolved = await Promise.resolve(actions(this.actionStack, this));
+    } else {
+      resolved = [].concat(actions);
+    }
     if (addToStack) {
       this.actionStack[this.currentID()] = actions;
     }
-    this.list = new List(actions, this.config);
+    this.list = new List(resolved, this.config);
     if (!this.selection) {
       this.initSelection();
     }