+++ packages/ui/src/lib/form/form.service.ts
@@ -28,7 +28,15 @@ export class FormService {
       .filter((field) => this.shouldBePartOfForm(field, form))
       .forEach((field) => {
         const validators = this.getValidators(field);
-        controls[field.property] = new FormControl(form.getValue(field.property), validators);
+        let value = form.getValue(field.property);
+        value = value === undefined ? null : value;
+        controls[field.property] = new FormControl(
+          {
+            value,
+            disabled: this.isReadOnly(field, form),
+          },
+          validators,
+        );
         // TODO use { updateOn: blur } when updating to angular 5.0.0
         // see https://github.com/angular/angular/commit/333a708bb632d4258ecb5fd4a0e86229fe9d26e4
       });
@@ -39,7 +47,7 @@ export class FormService {
   public addField(field: Field, form: Form<any>, group: FormGroup) {
     console.warn('addField is experimental!');
     const validators = this.getValidators(field);
-    const control = new FormControl(form.getValue(field.property), validators);
+    const control = new FormControl({ value: form.getValue(field.property) }, validators);
     group.addControl(field.property, control);
   }