+++ packages/ui/src/lib/io/input/input.component.ts
@@ -30,6 +30,8 @@ import { debounceTime } from 'rxjs/operators';
 export class InputComponent extends DynamicSlotComponent implements ControlValueAccessor, OnChanges {
   /** The belonging form group */
   @Input() group: FormGroup;
+  /** If true, the input is readonly. Is set by ControlValueAccessor#setDisabled */
+  disabled: FormGroup;
   /** The belonging form control. This is not required if you pass in a field and group. */
   @Input() control: AbstractControl;
   /** The changed ouput emits whenever the form control of the input changes. */
@@ -118,6 +120,7 @@ export class InputComponent extends DynamicSlotComponent implements ControlValue
       // console.warn('could not connect control: no instance loaded');
       return;
     }
+    this.componentInstance.disabled = this.disabled;
     if (this.componentInstance.registerOnChange && this.propagateChange) {
       this.componentInstance.registerOnChange(this.propagateChange);
     }
@@ -143,4 +146,15 @@ export class InputComponent extends DynamicSlotComponent implements ControlValue
   }
 
   registerOnTouched() {}
+
+  setDisabledState(isDisabled) {
+    this.disabled = isDisabled;
+    if (this.componentInstance) {
+      if (this.componentInstance.setDisabledState) {
+        this.componentInstance.setDisabledState(isDisabled);
+      } else {
+        this.componentInstance.disabled = isDisabled;
+      }
+    }
+  }
 }