+++ website/stackblitz/Stackblitz.js
@@ -1,0 +1,4 @@
+import React from 'react';
+import { Stackblitz } from 'react-stackblitz-snippet/index.jsx';
+
+export const StackblitzWrapper = (props) => <Stackblitz {...props}></Stackblitz>;